fix: corregir función ROUND en PARTE 3
ERROR [42883]: function round(double precision, integer) does not exist

Solución: Convertir a NUMERIC antes de ROUND
- ROUND(valor::FLOAT, 2) ❌
- ROUND(valor::NUMERIC, 2) ✅

Cambios:
- porcentaje_pagado: cast a NUMERIC
- dias_promedio_conciliacion: cast a NUMERIC antes de ROUND

Made-with: Cursor

+++ backend/sql/034_FASE2_PARTE3.sql
@@ -9,14 +9,14 @@ SELECT
     COUNT(*) as total_pagos,
     COUNT(CASE WHEN estado='PAGADO' THEN 1 END) as pagos_completados,
     ROUND(
-        COUNT(CASE WHEN estado='PAGADO' THEN 1 END)::FLOAT / NULLIF(COUNT(*), 0) * 100, 
+        (COUNT(CASE WHEN estado='PAGADO' THEN 1 END)::NUMERIC / NULLIF(COUNT(*), 0) * 100)::NUMERIC, 
         2
     ) as porcentaje_pagado,
     SUM(CASE WHEN estado='PAGADO' THEN monto_pagado ELSE 0 END) as monto_total_pagado,
     ROUND(
         AVG(CASE WHEN estado='PAGADO' AND fecha_conciliacion IS NOT NULL 
             THEN EXTRACT(DAY FROM fecha_conciliacion - fecha_pago) 
-            ELSE NULL END),
+            ELSE NULL END)::NUMERIC,
         2
     ) as dias_promedio_conciliacion,
     DATE_TRUNC('day', NOW()) as fecha_snapshot
